nekodigi/streamlit/requirements.txt
streamlit>=1.28
openai>=1.3
pillow>=10.0
